Fix file location and linefeeds
