+++ tests/PlantUml.Builder.Tests/ClassDiagrams/ClassEndTests.cs
