+++ tests/PlantUml.Builder.Tests/ClassDiagrams/ClassStartTests.cs
@@ -2,9 +2,8 @@ using System;
 using System.Text;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using PlantUml.Builder.ClassDiagrams;
 
-namespace PlantUml.Builder.Tests.ClassDiagrams
+namespace PlantUml.Builder.ClassDiagrams.Tests
 {
     [TestClass]
     public class ClassStartTests
