+++ tests/PlantUml.Builder.Tests/StringBuilderExtensions/InlineClassMemberTests.cs
